# contracts/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Quicker.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/audit/records/{entityType}/{entityId}": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Timeline of one record: every event with before/after values and the field-level diff",
        "operationId": "getAuditRecordsByEntityTypeByEntityId",
        "parameters": [
          {
            "name": "entityType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/collaboration/notifications": {
      "get": {
        "tags": [
          "Collaboration"
        ],
        "summary": "My notifications, newest first; pass the last id as 'before' to page",
        "operationId": "getCollaborationNotifications",
        "parameters": [
          {
            "name": "unreadOnly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationView"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/collaboration/attachments": {
      "post": {
        "tags": [
          "Collaboration"
        ],
        "summary": "Upload a file for a record (multipart/form-data: entityType, entityId, file)",
        "operationId": "postCollaborationAttachments",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentSummary"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Collaboration"
        ],
        "operationId": "getCollaborationAttachments",
        "parameters": [
          {
            "name": "entityType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collaboration/attachments/{attachmentId}": {
      "get": {
        "tags": [
          "Collaboration"
        ],
        "operationId": "getCollaborationAttachmentsByAttachmentId",
        "parameters": [
          {
            "name": "attachmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentSummary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Collaboration"
        ],
        "operationId": "deleteCollaborationAttachmentsByAttachmentId",
        "parameters": [
          {
            "name": "attachmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/collaboration/comments/{commentId}": {
      "put": {
        "tags": [
          "Collaboration"
        ],
        "operationId": "putCollaborationCommentsByCommentId",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Collaboration"
        ],
        "operationId": "deleteCollaborationCommentsByCommentId",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/collaboration/links/{linkId}": {
      "delete": {
        "tags": [
          "Collaboration"
        ],
        "operationId": "deleteCollaborationLinksByLinkId",
        "parameters": [
          {
            "name": "linkId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/collaboration/views/{viewId}": {
      "put": {
        "tags": [
          "Collaboration"
        ],
        "operationId": "putCollaborationViewsByViewId",
        "parameters": [
          {
            "name": "viewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedViewView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Collaboration"
        ],
        "operationId": "deleteCollaborationViewsByViewId",
        "parameters": [
          {
            "name": "viewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/collaboration/custom-fields/{fieldId}": {
      "get": {
        "tags": [
          "Collaboration"
        ],
        "operationId": "getCollaborationCustomFieldsByFieldId",
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomFieldView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Collaboration"
        ],
        "operationId": "putCollaborationCustomFieldsByFieldId",
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveCustomFieldRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomFieldView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Collaboration"
        ],
        "operationId": "deleteCollaborationCustomFieldsByFieldId",
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AttachmentSummary": {
        "required": [
          "id",
          "entityType",
          "entityId",
          "fileName",
          "contentType",
          "sizeBytes",
          "sha256",
          "uploadedBy",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "sha256": {
            "type": "string"
          },
          "uploadedBy": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CommentView": {
        "required": [
          "id",
          "entityType",
          "entityId",
          "authorMembershipId",
          "authorName",
          "body",
          "mentions",
          "parentId",
          "createdAt",
          "editedAt",
          "deletedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "authorMembershipId": {
            "type": "string",
            "format": "uuid"
          },
          "authorName": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "mentions": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "parentId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "editedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "deletedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "CustomFieldOption": {
        "required": [
          "value",
          "label"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "label": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "CustomFieldRules": {
        "type": "object",
        "properties": {
          "min": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "max": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxLength": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pattern": {
            "type": [
              "null",
              "string"
            ]
          },
          "referenceType": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Type-specific rules: number range, text length and pattern, the entity type a reference points at."
      },
      "CustomFieldView": {
        "required": [
          "id",
          "entityType",
          "key",
          "label",
          "description",
          "type",
          "required",
          "options",
          "rules",
          "indexed",
          "position",
          "active",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "entityType": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "label": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "description": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "type": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFieldOption"
            }
          },
          "rules": {
            "$ref": "#/components/schemas/CustomFieldRules"
          },
          "indexed": {
            "type": "boolean"
          },
          "position": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "EditCommentRequest": {
        "required": [
          "body"
        ],
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          },
          "mentions": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "JsonElement": { },
...
      "NotificationView": {
        "required": [
          "id",
          "kind",
          "title",
          "body",
          "link",
          "entityType",
          "entityId",
          "data",
          "actorMembershipId",
          "createdAt",
          "readAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "title": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "body": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "link": {
            "type": [
              "null",
              "string"
            ]
          },
          "entityType": {
            "type": [
              "null",
              "string"
            ]
          },
          "entityId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "data": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "actorMembershipId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "readAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "SaveCustomFieldRequest": {
        "required": [
          "entityType",
          "key",
          "label",
          "type"
        ],
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "label": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "type": {
            "type": "string"
          },
          "required": {
            "type": "boolean",
            "default": false
          },
          "options": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/CustomFieldOption"
            }
          },
          "rules": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CustomFieldRules"
              }
            ]
          },
          "indexed": {
            "type": "boolean",
            "default": false
          },
          "position": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 0
          },
          "active": {
            "type": "boolean",
            "default": true
          },
          "description": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "SavedViewView": {
        "required": [
          "id",
          "entityType",
          "name",
          "ownerMembershipId",
          "shared",
          "isDefault",
          "definition",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "entityType": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "ownerMembershipId": {
            "type": "string",
            "format": "uuid"
          },
          "shared": {
            "type": "boolean"
          },
          "isDefault": {
            "type": "boolean"
          },
          "definition": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SaveViewRequest": {
        "required": [
          "entityType",
          "name"
        ],
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "definition": { },
          "shared": {
            "type": "boolean",
            "default": false
          },
          "isDefault": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
    }
  },
...
}